Condition added to increase backwards compatibility (gif to png)
Added an extra line for here/rss.gif, now has a condition to check for existance of .gif and an extra line however for  the .png  variant. This should increase backwards compatibility for anyone who installs this version on Plone 4.2.5 (still has gif) however Plone 4.3 only has png and throws the error.